Move (Binary|Unary)Operator to enums
