+++ supriya/enums.py
@@ -17,6 +17,61 @@ class AddAction(IntEnumeration):
     REPLACE = 4
 
 
+class BinaryOperator(IntEnumeration):
+
+    ### CLASS VARIABLES ###
+
+    ABSOLUTE_DIFFERENCE = 38  # |a - b|
+    ADDITION = 0
+    AMCLIP = 40
+    ATAN2 = 22
+    BIT_AND = 14
+    BIT_OR = 15
+    BIT_XOR = 16
+    CLIP2 = 42
+    DIFFERENCE_OF_SQUARES = 34  # a*a - b*b
+    EQUAL = 6
+    EXCESS = 43
+    EXPRANDRANGE = 48
+    FLOAT_DIVISION = 4
+    FILL = 29
+    FIRST_ARG = 46
+    FOLD2 = 44
+    GREATEST_COMMON_DIVISOR = 18
+    GREATER_THAN_OR_EQUAL = 11
+    GREATER_THAN = 9
+    HYPOT = 23
+    HYPOTX = 24
+    INTEGER_DIVISION = 3
+    LEAST_COMMON_MULTIPLE = 17
+    LESS_THAN_OR_EQUAL = 10
+    LESS_THAN = 8
+    MAXIMUM = 13
+    MINIMUM = 12
+    MODULO = 5
+    MULTIPLICATION = 2
+    NOT_EQUAL = 7
+    POWER = 25
+    RANDRANGE = 47
+    RING1 = 30  # a * (b + 1) == a * b + a
+    RING2 = 31  # a * b + a + b
+    RING3 = 32  # a*a*b
+    RING4 = 33  # a*a*b - a*b*b
+    ROUND = 19
+    ROUND_UP = 20
+    SCALE_NEG = 41
+    SHIFT_LEFT = 26
+    SHIFT_RIGHT = 27
+    SQUARE_OF_DIFFERENCE = 37  # (a - b)^2
+    SQUARE_OF_SUM = 36  # (a + b)^2
+    SUBTRACTION = 1
+    SUM_OF_SQUARES = 35  # a*a + b*b
+    THRESHOLD = 39
+    TRUNCATION = 21
+    UNSIGNED_SHIFT = 28
+    WRAP2 = 45
+
+
 class CalculationRate(IntEnumeration):
     """
     An enumeration of scsynth calculation-rates.
@@ -331,3 +386,63 @@ class RequestName(StrictEnumeration):
     @property
     def request_id(self):
         return RequestId.from_expr(self.name)
+
+
+class UnaryOperator(IntEnumeration):
+
+    ### CLASS VARIABLES ###
+
+    ABSOLUTE_VALUE = 5
+    AMPLITUDE_TO_DB = 22
+    ARCCOS = 32
+    ARCSIN = 31
+    ARCTAN = 33
+    AS_FLOAT = 6
+    AS_INT = 7
+    BILINRAND = 40
+    BIT_NOT = 4
+    CEILING = 8
+    COIN = 44
+    COS = 29
+    COSH = 35
+    CUBED = 13
+    DB_TO_AMPLITUDE = 21
+    DIGIT_VALUE = 45
+    DISTORT = 42
+    EXPONENTIAL = 15
+    FLOOR = 9
+    FRACTIONAL_PART = 10
+    HZ_TO_MIDI = 18
+    HZ_TO_OCTAVE = 24
+    HANNING_WINDOW = 49
+    IS_NIL = 2
+    LINRAND = 39
+    LOG = 25
+    LOG10 = 27
+    LOG2 = 26
+    MIDI_TO_HZ = 17
+    SEMITONES_TO_RATIO = 19
+    NEGATIVE = 0
+    NOT = 1
+    NOT_NIL = 3
+    OCTAVE_TO_HZ = 23
+    RAMP = 52
+    RAND = 37
+    RAND2 = 38
+    RATIO_TO_SEMITONES = 20
+    RECIPROCAL = 16
+    RECTANGLE_WINDOW = 48
+    S_CURVE = 53
+    SIGN = 11
+    SILENCE = 46
+    SIN = 28
+    SINH = 34
+    SOFTCLIP = 43
+    SQUARE_ROOT = 14
+    SQUARED = 12
+    SUM3RAND = 41
+    TAN = 30
+    TANH = 36
+    THRU = 47
+    TRIANGLE_WINDOW = 51
+    WELCH_WINDOW = 50
