+++ supriya/synthdefs/SynthDef.py
@@ -7,6 +7,7 @@ import tempfile
 
 import yaml
 
+from supriya import BinaryOperator, UnaryOperator
 from supriya.realtime.ServerObjectProxy import ServerObjectProxy
 
 
@@ -214,12 +215,12 @@ class SynthDef(ServerObjectProxy):
             for ugen in self._ugens:
                 parts = [type(ugen).__name__]
                 if isinstance(ugen, supriya.ugens.BinaryOpUGen):
-                    ugen_op = supriya.synthdefs.BinaryOperator.from_expr(
+                    ugen_op = BinaryOperator.from_expr(
                         ugen.special_index
                     )
                     parts.append("(" + ugen_op.name + ")")
                 elif isinstance(ugen, supriya.ugens.UnaryOpUGen):
-                    ugen_op = supriya.synthdefs.UnaryOperator.from_expr(
+                    ugen_op = UnaryOperator.from_expr(
                         ugen.special_index
                     )
                     parts.append("(" + ugen_op.name + ")")
