--- supriya/synthdefs/UnaryOperator.py
@@ -1,61 +1,0 @@
-from uqbar.enums import IntEnumeration
-
-
-class UnaryOperator(IntEnumeration):
-
-    ### CLASS VARIABLES ###
-
-    ABSOLUTE_VALUE = 5
-    AMPLITUDE_TO_DB = 22
-    ARCCOS = 32
-    ARCSIN = 31
-    ARCTAN = 33
-    AS_FLOAT = 6
-    AS_INT = 7
-    BILINRAND = 40
-    BIT_NOT = 4
-    CEILING = 8
-    COIN = 44
-    COS = 29
-    COSH = 35
-    CUBED = 13
-    DB_TO_AMPLITUDE = 21
-    DIGIT_VALUE = 45
-    DISTORT = 42
-    EXPONENTIAL = 15
-    FLOOR = 9
-    FRACTIONAL_PART = 10
-    HZ_TO_MIDI = 18
-    HZ_TO_OCTAVE = 24
-    HANNING_WINDOW = 49
-    IS_NIL = 2
-    LINRAND = 39
-    LOG = 25
-    LOG10 = 27
-    LOG2 = 26
-    MIDI_TO_HZ = 17
-    SEMITONES_TO_RATIO = 19
-    NEGATIVE = 0
-    NOT = 1
-    NOT_NIL = 3
-    OCTAVE_TO_HZ = 23
-    RAMP = 52
-    RAND = 37
-    RAND2 = 38
-    RATIO_TO_SEMITONES = 20
-    RECIPROCAL = 16
-    RECTANGLE_WINDOW = 48
-    S_CURVE = 53
-    SIGN = 11
-    SILENCE = 46
-    SIN = 28
-    SINH = 34
-    SOFTCLIP = 43
-    SQUARE_ROOT = 14
-    SQUARED = 12
-    SUM3RAND = 41
-    TAN = 30
-    TANH = 36
-    THRU = 47
-    TRIANGLE_WINDOW = 51
-    WELCH_WINDOW = 50
